Add files via upload
The "SRC" folder is consisting of the various other sub folders such as components for flag, hooks for the flags and the main page.tsx. The Public folder is having the JSON file and the image file, flag.png. There is also an img file which is the screenshot of the app running on the local host.